+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Added
 
 - [core] Generate node and deployment UUID
+- [core] Return UUIDS after setup
 
 ## Changed